Support reading DNS settings from NetworkManager on Linux
Support reading DNS settings from NetworkManager on Linux. Call
nmcli-tool to get the DNS settings.

On Ubuntu 12.04, NetworkManager is setup by default to run dnsmasq on
127.0.0.1:53. This causes reading the DNS server from /etc/resolv.conf
to return that address, which will not work as an address for
webpage-replay to use.

Does not support setting the DNS, as dnsmasq is most likely listening on
127.0.0.1:53. Thus, this fixes problems only for --no-dns_forwarding
use-cases.